SBM17175: Optimize aplay & arec for swiapp using
Aplay process and Arec process are started and killed by swiapp at present.
For example, swiapp will start aplay when playing wav, and kill aplay when playing done.
This call method is too rough and waste of time, even fail to kill aplay and arec sometime.
So we would like to make swiapp to call functions in aplay and arec directly, i.e. make aplay and arec as dyncmic libraries.
This patch has been tested with the latest swiapp.

Change-Id: I0618c788ed1c1cf6290a1bae4d0a344b0fa33cd5
Signed-off-by: ykong <ykong@sierrawireless.com>