trac:2488: In order for smackfsdef='*' to be accepted as mount option, running process must have '*' label as well. I am not entirely clear on why would this be the case (there is nothing documented out there), but it currently works, and investigation is under way. There are indications from the community discussion threads, that indeed something is not quite right when using smack mount options, and how and when these really apply.
smackfs must have '_' label. It may not be necessary to
add this label explicitly, but it is here to signify its
 importance in this case.

Mounting of smackfs is not necessary, but it is a nice-to-have
add-on in case someone needs smack kernel interface. Changes
in mdev.conf are necessary to allow proper Legato operation.

Change-Id: Id46d3a054c78588edd8f3f26165cb1f67517f51e