ticket:3105 - Improve start-up time : Suppress kernel messages output
LK set the "quiet" option to the command line passed to kernel
This "optimization" may be suppressed by setting the custom field
	AT!CUSTOM="BOOTQUIETDISABLE",1
	AT!RESET
If set to 1, the kernel messages will be output to the console.
If not set (default), the kernel message are not output to the console.

Patch 3: Updated after review comments.

Patch 2: Check for "BOOTQUIETDISABLE" custom field to add the "quiet"
         option on the command line (NEW83465).

Patch 1: Original

Change-Id: If30900d365b3d665b59180b724b20d6e7a043cde