trac:3234 - Do not open standard FDs in Sierra Wireless init
If there is no console available, this init will keep on respawning.
In order to avoid this situation, we would need to emulate/copy
certain things from "real" init, and that would not be optimal.

There isn't really great deal of things happening in this script
anyhow, and having no console output for short period of time should
not be a big deal (BB init will initialize standard FDs once it takes
over).

stdin/out/err could always be added back to aid debugging if needed,
and when console is available.

Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Change-Id: I294d145bbc593904ecc43a2f97c242f621fcfac0