trac:4047 - Detect squashfs volume at run time
The ramdisk script for mountint rootfs assumed that if a UBI
magic ('UBI#') was found in a mtd partition, then the whole
partition is assumed to be ubifs. It then proceded with mounting
a ubifs file system.

When support for squashfs UBI volumes was added, partitions were
properly detected as UBI partitions, but the mount failed due to
file system mismatch. Another identification step was needed to
detect a squashfs volume.

Read the attached ubi0_0 device to extract the file system magic.
If the detected magic corresponds to squashfs ('hsqs'), attach a
UBI volume and mount squashfs file system from the volume.
Otherwise assume ubifs and mount ubifs device.

Change-Id: Ic368f1ec5d456665bab7225f2fa613f84f619387
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>