[9xxx] Optimize linux-quic compilation
There are two ongoing issues with mdm9x40 kernel compilation:
a. the source directory is also used as a build directory
b. incremental builds are not supported

As to solve this problem:
a. build directory is set to "${WORKDIR}/build"
b. ${SRC_URI} is added to support incremental compiling

This also synchronize mdm9x28 and mdm9x40 build recipes.

Resolves: DEV90797
Change-Id: I90640a46d9f3a20d301af9d88b9a7257f9654759
Signed-off-by: Yan Kong <ykong@sierrawireless.com>
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>