[9xxx] Remove legato-af from rootfs
Libraries packaged in applications, such as libqmiservices for modem
services, where messing up with the shlib dependency check of Yocto.
It was falsy considering legato-af as the provider for this libraries,
which ended up making other qmi packages falsy dependend on Legato.

Since Legato is now in a separate partition also on these devices,
this removes the legato files from the rootfs, which also solves the
problem mentioned above as the Yocto packaging subsystem doesnt
encounter these libraries anymore.

Change-Id: Ia7c669b9ac336656671d126f50ea3e6903e6d3db
Resolves: trac:4671 trac:4672