Prevent dnsmasq from not starting in case of invalid symlinks
As to prevent wear levelling, some applications are managing the content
of the dnsmasq configuration through a symlink from /etc/dnsmasq.d to
/tmp/<something>.

Since the content of /tmp is volatile, the symlink is not guarranted to
be valid on boot, which prevents dnsmasq from starting.

If it finds an invalid symlink before starting dnsmasq, the init script
will create an empty file at the destination to make the symlink valid.

Resolves: trac:4829
Change-Id: Ic3620d8db9a563c63075d0ff3e2df45df6203d19