Update default /etc/network/interfaces file
The default file provided in poky layer mention interfaces such
as eth1 and wlan0 which either doesn't exist or are not controlled
through this file.

It also provide defaults for usb0 as well, which is actually being
managed by a Legato tool called configEcm.

On mdm9x15 there is already a file provided for the wifi, which is
barely updated, while a new file (with lower priority since the common
layer is lower priority) is being introduced in the common/ layer.

Resolves: trac:4790
Change-Id: Iea627b4944f39d313f6464c3c16f7bb185160360