uSD cards are not recognized even available at power up
At the moment, hot plugging of the uSD cards on WP85 platform
does not work, and having msm_sdcc module loaded using
kernel hotplug or manually does not work either.

In order to work around this problem, we could:
 - Add CONFIG_MMC_MSM=y to kernel config file
 - Add msm_sdcc to modules file

In order to keep monolithic kernel size in check, it is best
to solve this problem by adding kernel module to modules file.

Resolves: trac:5072
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Change-Id: Ie9943697548222a8939742555aaf5338216561e4