[AR758X] Set Kernel and rootfs image to 4k page size as Default
In the coming days, it will only use 4K page Nand flash device. As
require we should set the image to 4K page as default for UBI and
kernel image. For the old device with 2k page size, LK will auto
detect and switch the image from 4K to 2K.

Resolves: QTI9X28-243
Change-Id: I4704bbd367aa7eefbf492dc44aac17f36850afd8
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>