[9x40] Fix syslogd startup script
When cryptsetup recipe was added, default shell ended up changing from
"bin/sh -> /bin/busybox.nosuid" to "bin/sh -> /bin/bash".
Script "/etc/init.d/syslog" is using keyword "local" out of a function,
which is a syntax error that is not encountered with busybox but doesn't
work with bash.

Therefore, this fixes the execution of the syslog init script with bash.

Resolves: QTI9X40-571
Change-Id: I1e39b851321b4c030b7dc2ed8471d373c86f4a72
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>