Update Legato recipe to build from work directory
The legato-af recipe is currently handling multiple cases for
fetching the legato framework, and it defaults to GitHub.
Since the project is now scattered into multiple repositories, with
an integration relying on repo, managing that case would end-up
making the recipe a lot more complex.

Since there were already some complaints about the complexity of this
recipe, this changes the behavior as to not take care of the fetching
of Legato at all.

It is left to the person (or machine) handling the Yocto sources to
place the Legato sources in a folder, that can either be called legato/
an placed at the same level as the Yocto layers (meta-swi, ...)
OR
that can be anywhere in the filesystem, as long as the variable LEGATO_WORKDIR
points to it.

Also:
- Some code in legato-tools was only relevent for Legato < 15.07.0.Beta so
  it has been removed.
- All source files from the LEGATO_WORKDIR are added to the SRC_URI
  variable so that bitbake can track changes on this files and
  re-trigger the legato-tools/legato-af build

Resolves: LXSWI9X1517-96 LXSWI9X28-8 LXSWI9X40-23
Change-Id: I0c6faecfb7702d6a028cd15e8efebcb8b4a5ef2c