Update LE.3.1 alx driver
The alx driver used by MDM9x40 is out of tree Kernel driver, which is
pulled from Code Aurora before compiling it.

Currently the LE.3.1 alx driver has two issues:
a) compile error
b) cause module watchdog to bark

For a), fix it by 0001-Fix-the-compile-error-in-LE.3.1.patch.
For b), now remove new codes from LE.3.1 and fallback to LE.2.0 currently.
Will fix it later when we get the solution form Qualcomm.

Resolves: QTI9X40-445
Change-Id: I2ba55f453c4b38028f0ef79f46b50b3827c58a81