[ecm] Add option to bridge ECM to Ethernet
WHQL/HLK certification test consists of two PCs connected via USB
modules to an Ethernet/IP backbone. Client PCs generate/receive
traffic whereas server PC relays traffic back to the client(s).

To support this test, add a script that bridges ECM to Ethernet
interface so that the two PCs to which the modules are attached
could communicate over Ethernet. This script is not enabled by
default and hence is not symlinked in /etc/rcS.d. To enable this
configuration, one has to do the following:

1. Descend into init script directory:
    $ cd /etc/rcS.d

2. Disable ECM configuration script:
    $ mv S95bringup_ecm.sh _S95bringup_ecm.sh

3. Re-link the ECM startup script to this new file:
    $ ln -s ../init.d/bridge_ecm.sh S95bringup_ecm.sh

This setting would be persistent across reboots.

The inverse should be done to restore normal operation.

Resolves: LE-5403
Change-Id: Ie04b88cf3c67fc0e786d7e0fb63031dfaa2e08a1
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>