Provide LEGATO_SYSROOT in legato class
If LEGATO_SYSROOT is not defined, the --sysroot option is not provided to gcc
resulting in the compiler using the default embedded sysroot which is tied to the
initial build directory and might be invalid.
This is especially the case when relying on the Yocto sstate cache.

The feature available has been made available in Legato,
but there is no mention of Yocto the CC env is not used by Legato.
To fix this issue, this defines LEGATO_SYSROOT in legato.bbclass.

(cherry picked from commit 903df0138457848d74f746d810689f9e62f18c5b)

Resolves: CORONADO-114
Change-Id: I450695671c4c42e6330c42a1fd9548b075880f03
Signed-off-by: Yoking Wang <yowang@sierrawireless.com>
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>