Bring up alx driver on the Qualcomm LE.3.1.rb.1.1 stack
This commit includes the following issues:
a. Fix Makefile to generate alx.ko.
b. The usage of IPA interface in alx driver doesn't match LE.3.1 IPA module.
   Remove the marco MDM_PLATFORM to disable IPA on alx driver, and fix the
   compile error when removing MDM_PLATFORM.

Note:
This commit also removes the patches in the following commits:
3f34bb0 Require IPA RM resource when wakeup module or plugin Ethernet Cable
9780480 Update LE.3.1 alx driver
This is because the IPA feature is disabled in this commits. It doesn't need to
apply the patches in 3f34bb0 and 9780480.
We will apply the Qualcomm CRs in the future when they fix the IPA issue.

Resolves: QTI9X40-445
Change-Id: Ia1b25bc3b8d8f06fb5f29a032a6c61e3c6bd7ed2
Signed-off-by: Yan Kong <ykong@sierrawireless.com>