[9x28] Dm-verity
The purpose of dm-verity is to implement a device mapper target
capable of validating the data blocks contained in a file system
against a list of cryptographic hash values.

We can close this function by set it to "DM_VERITY_ENCRYPT = off"
in file "swi-mdm9x28.conf". The root hash of the rootfs is stored
in scripts in initramfs, So when dm-verity is enabled we must burn
the kernel and the rootfs together, they are pairs or you will got
crash during boot up.

Dm-verity is actived from initramfs. Till now only the rootfs partition
is protected by Dm-verity.The rootfs data is stored at UBI volume 0 (UBI0_0)
and the corresponding data hash tree table is store at volume 1 (UBI0_1)
Every access to the data blocks (mtdblock0_0) Dm-verity will check
corrsponding hash tree table (mtdblock0_1). If the data and the hash is
not match then return I/O error, which is expected.

notes: this is merge from QTI9X40-471

Resolves: QTI9X28-420
Change-Id: Iceb0bef8f1055e9f8d971628d75211d97b24da15
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>