Add mangOH i2c utility i2cgpioctl for mdm9x28
The i2cgpioctl utility provides a simple testing interface for mangOH
peripherals on the external i2c bus.  It also supports probing for
internal i2c devices, since the bus is shared.  This change pulls in
the WP8548/WP75xx version of the tool, and adds flexibility so that
the i2c bus can be specified by the user.  Minor improvements are
also made for code cleanup.

Resolves: LXSWI9X28-15
Change-Id: I758a58f1a267448286409590e9c322fccacb686b
Signed-off-by: Ryan Kirk <rkirk@sierrawireless.com>