[AR759x] Support AT!USBVID and AT!USBPID
Description:
Currently, the usb vid and pid are configured by "composition-sierra_dev" script.
The usb vid and pid could not be changed. Change the way for getting the vid and pid.
Use the UD module for getting the usb vid and pid. The UD module will get the use
vid and pid from the sharing memory which contains the vid and pid during boot.

Root cause:
N/A

Solution:
- Bring up UD module
- Implement two AT commands to change/query USB VID/PID

Resolves: QTI9X40-959
Change-Id: I935c338b85631dca072f42a83133b6460305083f
Signed-off-by: liliu <liliu@sierrawireless.com>