legato: fix tinycbor build dependency
Yocto builds number of packages in parallel. Consequently, there could
be race condition between packages, if dependencies are not manually
specified.

In this case, legato depends on development portion of tinycbor package.
However, tinycbor package output may not be available at the time of
legato build.

To fix this, we need to explicitly specify that legato depends on
tinycbor, and it needs to be built first.

Resolves: Jira:LXSWI9X1517-161
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: Ieea2cfe247d1bf01844926f3c6dc5e15ba2e2857