Implement "USBCOMP" for USB composition
There are three ERD items about USB composition. They are respectively
1. The module shall support interface bitmask to enable/disable interfaces.
2. The module shall provide non-volatile memory to store USB composition info.
3. The module shall support AT command to set/query/test USB composition.

Currently, the USB ports and functions are configured by "composition-sierra_dev"
script. The USB ports and functions are not changed. So modify the script to
get the ports and functions by "ud_getusbinfo" app. The "ud_getusbinfo" app
will get the information about the ports and functions from the sharing memory
which contains it during boot.

Resolves: QTI9X40-507
Change-Id: I89b557b06686984890ada8fbb9114a10fb2d520b
Signed-off-by: liliu <liliu@sierrawireless.com>