Move the root hash of rootfs-squashfs to UBI volume of rootfs-UBI
Description:
Currently the root hash of rootfs.squashfs for DM verity is stored
in Ramdisk of kernel. There is a side effect for this design: we
have to download the kernel and Rootfs together during update. Now
move this root hash to one of the UBI volume of rootfs-UBI image.

Resolves: QTI9X40-1111
Change-Id: I2ee15bafc1c45381cfa054f8d5ec91c1e91190a8
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>