Move the root hash of rootfs-squashfs into UBI volume of rootfs-UBI image
Description:
Currently the root hash of rootfs.squashfs for DM verity is stored
in Ramdisk. There is a side effect for this design: we have to download
the kernel and rootfs together during update. Now
move this root hash into one of the UBI volume of rootfs-UBI image.

Note: This is merged from QTI9X40-1111 and QTI9X40-1196.

Resolves: QTI9X28-708
Change-Id: I29e4da057efa2fade0b003126c53b4361f0050ab
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>