Do not fudge rootfs as default
We use "fudge_ro_rootfs=true" in kernel cmdline to control if we need
to fudge RO rootfs. If we has "fudge_ro_rootfs=true" in our kernel cmdline,
partition "customer2" will be mounted as UBIFS, and /etc, /data, e.g.
will overlay on this partition. This param is for debug purpose, and we
don't use it as default.

Resolves: QTI9X40-1006
Change-Id: Ib3a857ffb170ca4698d76b63876dea572a2f0b2e
Signed-off-by: rli <rli@sierrawireless.com>