Do not fail on binaries from Legato already stripped
Since 4acb6642b was merged on Legato, binaries generated by Legato
are already stripped as part of the build process.

Yocto generally strip binaries as part of the do_package process,
as to store debug symbols is a -dbg package that can be used for
debugging without putting the (heavy) debug symbols in the rootfs.
In the Legato case, Yocto is complaining since it can't collect these
debug symbols, and therefore stops the build for the legato-af package.

Resolves: LXSWI9X28-17
Change-Id: I182e5a38a8d4986cbc4eb3de6aa90c44a7271acf