Do not fudge rootfs as default
We use "fudge_ro_rootfs=true" in kernel cmdline to control if we need
to fudge RO rootfs. If we has "fudge_ro_rootfs=true" in our kernel cmdline,
partition "customer2" will be mounted as UBIFS, and /etc, /data, e.g.
will overlay on this partition. This param is for debug purpose, and we
don't use it as default.

Porting from Rock's commitment, which commit id is 66af88 in this branch.

Resolves: QTI9X28-860
Change-Id: Id8678cd2b0d3ef61f9405318e0f54f4e14ff4c44
Signed-off-by: Lucy He <LHe@sierrawireless.com>