Change networking script startup and shutdown orders on MDM9x40
MDM9x40 supports IPA feature. Network interface cannot work if IPA
module is not ready when enabling IPA feature.

The original startup order of networking is 20. The orders of
IPA startup scripts are 30+.

So network cannot work when enabling IPA feature on MDM9x40. Now
change startup order of networking script to 40, also change the
shutdown order of networking to 60.

Resolves: QTI9X40-1191
Change-Id: If2140ddd7f5e15918e918405e68d62b18f308307
Signed-off-by: Yan Kong <ykong@sierrawireless.com>