Update legato dependency on flash tools
The legato-af package needs squashfs/ubi/yaffs2 tools at build time
and they were missing from the dependenciees, so they should be added.

The legato-image package already has these dependencies, but they are
for the target version of the package while the host version are needed
as well, so this should be fixed.

Resolves: LXSWI9X1517-201
Change-Id: I400f8b3ef734d6dc335d627649fdf57838a33e66