Add UART1 Linux console support in 9x40
For mdm9x40, Currenlty UART2 is as the default console. But we also need
to support UART1 as console. Currently UART2 is enumurated as ttyHSL0.
We need to change UART1 as ttyHSL0, and UART2 as ttyHSL1.

So we need to add ttyHSL0 and ttyHSL1 in the Linux command line. And
select the actual console with the AT command setting.

(cherry picked from commit a50668e06f35b66b1e945cd769e2fc12fc7beacf)

Resolves: QTI9X40-1244
Change-Id: I0a12eb9ed90a54e45ef9315f2268b2af403ad815
Signed-off-by: Kinbo Huang <KHuang@sierrawireless.com>
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>