Fix failture to mount rootfs occasionally during boot up stress test
Description:
During power on off stress test there is a very low probability that
can not find the squashfs symbol "hsqs" and then mount fail.

Root cause:
The devices of UBI volume are not generated once "ubiattach" is executed
successfully,e.g. ubi0_4 is ready but ubi0_2 or ubi0_3 is not ready (low
probability).

This is a known issue for ubiattach, Detail is mentioned in file:
"mtd-utils/tests/ubi-tests/README.udv":
There is a problem with udev: when a volume is created, there is a delay
before corresponding /dev/ubiX_Y device node is created by udev, so some
tests fail because of this. The symptom is error messages like
"cannot open /dev/ubi0_0".

Solution:
Here is a work-around for the issue:
Need to check all the necessary devices of UBI volume after "ubiattach"
is executed successfully, if devices not ready wait it a short time.

Resolves: QTI9X28-967
Change-Id: I9c4f7878c6efac1679daf3c162505511b19f62e7
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>