Remove gdbserver from rootfs and ship it with toolchain
gdbserver being GPLv3, it is a bit of an anomaly on the rootfs
which causes licensing issues.
This removes it from the rootfs, but builds it and ship it anyway
in the toolchain, that way we can eventually transfer it to the
target to provide remote debugging anyway.

At this time the removal is only for the mdm9x15 devices but will
be extended to other platforms.

(cherry picked from commit d8aef8c9c0f18b7494f2351c7258fe8acdffb3e0)

Resolves: LXSWI9X28-26
Related: LE-4043 LXSWI9X1517-209
Change-Id: I733faac66817e13e9693d65cb384840f79368970
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>