Fix glib compilation under gcc 6.2
Glib relies on some gcc checks against non-literal strings being used as
format strings in printf-like functions. But it needs to grant itself
some exceptions in a few places using a #pragma.  The story seems to be
that gcc has extended the checking to functions which are like strftime,
and glib needs one more such exception in one module where it calls
strftime with a computed string.

* common/recipes-devtools/glib/glib-2.0_2.40.0.bbappend: New file.

* common/recipes-devtools/glib/files/compile-glib2-with-gcc-6.2.patch:
New file.

Resolves: LE-6350
Change-Id: Ifdcd7a51bd2e8d6fd9153e2f71217501ba4d97fc