Add cryptsetup-native dependency to legato-af recipe to fix Yocto build issue
Description:
This tool "veritysetup" is used to create the image's hash table
for Dm-verity. Legato image in this branch is working with
Dm-verity, so need to add the dependency for legato build or it
will build fail. The Error MSG like this: "Unable to find
veritysetup within toolchain, please update your toolchain".

Root cause:
"veritysetup" tools is not created during build.

Solution:
Add veritysetup tools dependency to legato bb.

Resolves: LXSWI9X40-49
Change-Id: I35f9a40fda9c7f336bd5c4e0e3a6a5e9f796be69
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>