Disable adbd
To prevent startup of adbd daemon if it is not
enabled in AT!USBCOMP setting.

Solution:
Check to start adbd is put based on values
saved in:
/sys/class/android_usb/android0/functions
if "ffs" is present, then it is enabled else not.
Also, now adbd is started from symbolic link in
/etc/rcS.d/ only. Earlier, it was getting
started from composition-sierra_dev file as well.

Resolves: QTI9X07-125
Change-Id: I68da9ba368bb342e1774ec88b7c39aa607b2813f
Signed-off-by: poonam infosys <pinfosys@sierrawireless.com>