Build tinycbor so lib with LDFLAGS
There is a new check on Yocto 2.2 regarding the presence of GNU HASH.

At the moment the .so lib from tinycbor is not generated using LDFLAGS
provided by Yocto, resulting in a QA error.

For some reason the CFLAGS value as declared in the Makefile is not
being used as well.

Resolves: LXSWIREF-7
Change-Id: If59e8c2c09fc275c7176911f3049d7dba54d2a8f