Fix dependencies of linux-quic on some native utils.
* meta-swi-mdm9x50/recipes-kernel/linux/linux-quic_git.bb
(DEPENDS): Variable append removed: not working.
(do_deploy[depends]): New append. Here we make the do_deploy
step for linux-quic depend on the dtbtool and mkbootimg
being installed into sysroot. This was suggested by
Bertrand. It seems to be working. If I do "cleansstate"
on these tools, and run the deploy step of linux-quick,
it's building the tools.

* meta-swi-mdm9x40/recipes-kernel/linux/linux-quic_git.bb: Likewise.

* meta-swi-mdm9x28/recipes-kernel/linux/linux-quic_git.bb: Likewise.

Resolves: LXSWIREF-18
Change-Id: I5aeca443aa7f71bc827811dc344ede83e1879b54