localgit.bbclass: allow do_fetch.
It's useful to be able to to specify SRC_URI material
and have it fetched in recipes that inherit from localgit
(i.e. which do not fetch their main source code but get
it via a symlink).

* common/classes/localgit.bbclass (do_fetch): Remove
empty function so do_fetch is not overriden to a no-op.
(do_unpack): Remove the function, so the symlink creation
is not done in do_unpack.
(do_fetch_prepend): Create the symlink here, before the
fetching. This requires Python code. The reason we want
the symlink created before the fetching is so the fetching
can go through the symlink, if desired.
It has to be done here rather than do_unpack, because
in do_unpack, the removal of ${PN}/ to make way for the
${PN} symlink will remove fetched files!

Resolves: LXSWIREF-6
Change-Id: Id960f6224c148456d530cf4490f96bbebe0e6a63