mdm9x50: Move libhardware.bbappend to meta-swi-extras.
This anticipates an upcoming change to switch libhardware
to use the implementation from ${WORKSPACE} rather than
codeaurora on the mdm9x50.

Since mdm9x50 won't be using the libhardware.bb in this
tree, we have to solve the problem of how not to include
that recipe when building mdm9x50, without replicating
the base definitions for the mdm9x28 and mdm9x40.

We can address the problem by moving the base definitions
to a bbclass, and turning the mdm9x28 and mdm9x40
bbappend files into bb files, which inherit from
that class. For the mdm9x50, there is no thus no
libhardware.bb, so the libhardware recipe is not activated
in this tree for that platform.

* meta-swi-mdm9xxx/conf/layer.conf (META_SWI_MDM9XXX_LAYERDIR):
New variable; provides path to meta-swi-mdm9xxx layer.
We need this inside the libhardware.bbclass to locate
autotools.patch, which is not moving anywhere.

* meta-swi-mdm9x50/recipes-bsp/libhardware/libhardware.bbappend:
File removed.

* meta-swi-mdm9xxx/recipes-bsp/libhardware/libhardware.bb:
Moved to common/classes/libhardware.bbclass

* meta-swi-mdm9x28/recipes-bsp/libhardware/libhardware.bbappend:
Renamed to libhardware.bb
(inherit): New inherit line incorporates libhardware
base material that was previously in libhardware.bb
from libhardware.bbclass.
(FILESEXTRAPATHS_append): Point to the libhardware files
in the meta-swi-mdm9xxx layer so the autotools.patch
can be found.

* meta-swi-mdm9x40/recipes-bsp/libhardware/libhardware.bbappend:
Likewise.

* meta-swi-mdm9x15/recipes-bsp/libhardware/libhardware.bb: New file.
Previously mdm9x15 didn't have a .bbappend; it has nothing
to override in the mdm9xxx libhardware.bb file. Now we need
trivial .bb file, otherwise mdm9x15 has no libhardware.

Resolves: LXSWIREF-6
Change-Id: I936b234633038cf25069583c1046cbce574f5429