mdm9x50: fix funny path handling in gen_master_dtb.
I ran into a build failure with the cat $kernel_img ..
step not finding the dtb file. This is mainly due to
the $dtb_files list being shored up using a find
over arch/arm/boot/dts, but then stripped down to
just basenames which are assumed to exist specifically
under arch/arm/boot/dts/qcom.

What we do is simplify the find job to keep the complete
path names, with .dtb suffix and all. Then inside the
loop which iterates over them, we do the trimming to
get the value that is interpolated into the dtb output
name.

I think the original

  targets=$(echo ${d#${BASEMACHINE_QCOM}-})

is a bug. Some of the file names that contain
$BASEMACHINE_QCOM do not contain it as a prefix.
The new logic removes the substring which
matches ${BASEMACHINE_QCOM}- from the names.

* meta-swi-mdm9x50/recipes-kernel/linux/linux-quic_git.bb
(gen_master_dtb): Add sanity check for existence of
BASEMACHINE_QCOM and alter the logic as discussed
above.

Resolves: LXSWIREF-6
Change-Id: I82871c52d15c3f9cc710f19ba9c4f5313d0a68fc