iptables: Allow platform-specific ECM names
ECM interface may be named "usb0" or "ecm0", or maybe "eem0". This
is currently dependent on how the gadget function is named in each
kernel version. On the other hand, iptables-restore script does
not allow any environment variable substitution, which makes them
platform-specific (need to name ECM interface per-platform).

The solution is to use an ECM_IF macro and subustitute it with
dynamically-determined ECM interface name before loading the file
to iptables-restore (or ip6tables-restore for IPv6).

Resolves: LXSWIREF-38
Change-Id: I2370d07a422959506e282d027d2e61fb55a29f4e
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>