networking: Initialize eth0 when no ECM is present
When no ECM is present, an attempt to retrieve ECM interface name
would return error, which in turn causes network configuration to
abort. In this case, no iptables rules would be applied and none
of network interfaces - with exception of lo - would be brought
up.

Ignore return code from interface search if ECM is not present.
Also in this case ignore iptables rules for ECM.

Resolves: LXSWIREF-55
Change-Id: I4295222a68fcf53eb9290dff9386f8b0be294faa
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>