mdev: -s recurses into /sys/devices to build /dev/bus.
This contains two new patches.

One patch causes "mkdev -s" to include the
/sys/devices tree. The side effect of this is that it will
nicely build up those /dev/bus/usb/XXX/YYY entries,
using existing logic (we don't have to touch any of the
device node making functionality).
The /sys/devices tree is a deeper than the hard-coded
maximum depth, so a tiny bit of refactoring is done to
gain a little dynamic flexibility in this area.
Symlinks under /sys/devices are not followed; this is
not necessary, and does cause time-consuming looping.

The second patch implements the d_type optimization
in BusyBox's recursive_action function for directory
traversal. This is because with the previous patch,
the "mdev -s" time went from 0.38 seconds to 1.12
seconds. With this optimization, the time is around
0.32 seconds: faster than originally without the
/sys/devices traversal!

* common/recipes-core/busybox/busybox_%.bbappend (SRC_URI): Include new
patches, mdev-dev-bus-usb.patch and recursive-action-optimize.patch.

* common/recipes-core/busybox/files/mdev-dev-bus-usb.patch: New file.
The change details are in the patch header.

common/recipes-core/busybox/files/recursive-action-optimize.patch:
Likewise.

Resolves: LXSWIREF-60
Change-Id: If8231bfdcbb4517366e26b3574a3e264fb469e8a