Fix the linux-quic do_config: missing linux-quic symlink.
This is the main issue reported in LXSWIREF-91.
However, the linux-quic recipe still doesn't completely
build if only this is fixed.

What is is happening is that the linux-quic symlink to the
git repo is being removed and replaced by an empty directory.

What is doing this is the the processing of the 'cleandirs'
flags variable attached to the do_unpack action.

* common/classes/localgit.bbclass (do_fetch_prepend):
Renamed to do_fetch: we don't want to prepend but take
over the entire action. Also, since this is python,
it should have the python keyword.
(do_unpack, do_patch): Blank actions to suppress anything
coming from the recipes or base bbclass. A localgit
recipe doesn't require any unpacking or patching.
(python): Global python code snippet to blank out
the cleandirs flag variable of the do_unpack action.
do_unpack still executes if we blank ot out, and its
cleandirs is still processed.

Resolves: LXSWIREF-91
Change-Id: I97779756c55c574c454c503765ce12c6a9f04def