i2c-tools: use alternatives not clash with busybox.
I discovered in the output of mdm9x28-image-minimal.bb:do_rootfs
that busybox is trying to make symlinks for i2c-tools, but
failing because they are real executables from another
package. We can Yocto's own ALTERNATIVE mechanism to properly
implement this.

* common/recipes-devtools/i2c-tools/i2c-tools_%.bbappend: New file.
What we do is redirect i2c-tools to install the clashing executables
as /usr/bin/<program>.i2c-tools and for each one of these assert
it as an alternative, vying for the symlink /usr/bin/<program>,
much like what Busybox's recipe is doing (that recipe does
it with under-handed Python calls).

The ALTERNATIVE_PRIORITY is then set so that i2c-tools wins and
gets the symlinks.

Resolves: LXSWIREF-91
Change-Id: Ic790e101e831f206407c24e01d517cb26a48ea54