reverting: Fix management of host ECM mac address on wp76
During tests it was discovered that the host may decide to use a
different MAC address from the one offered by the gadget. In that
case the dhcp server - once it finds that the only available
address for ECM (which is by default 192.168.2.3) is already
reserved - would deny any further DHCP requests.

Reverting this patch may cause delayed address assignment for
clients that initiate the session with DHCPREQUEST (rather than
DHCPDISCOVER).

Resolves: LE-6265
Related: LXSWIREF-38
Change-Id: Ic4f5486da9ea100c465a71dd76731941bba89cbb
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>