iptables: Copy ecm firewall rules to bridge0
This is a temporary fix to allow bridge0 to act as a surrogate ecm
interface on platforms that (forcefully) enslave ecm to bridge0.
The fix allows that all requests honoured on ecm also get honoured
on bridge0, hence all ssh, ping and dhcp access will be granted.
For platforms not having bridge0, these rules have no effect.

Also fixed typo in one of the rules that changes sequence.

Resolves: LXSWIREF-38
Related: LE-6265
Change-Id: I22cd9f90034f31442a549c7bee3aea4cf5bae199
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>