mdm9x15: build kernel 3.14 under Yocto 2.2.
* meta-swi-mdm9x15/recipes-kernel/linux/linux-yocto-rt_3.14.bbappend:
This file is renamed to .bb because there is no linux-3.14 in poky
which can take the bbappend. We include the "linux-yocto.inc"
from poky which has most of the implementation, similarly to what
the poky linux-4 recipes are doing.
(PV): We need to set this version variable, because there is some
version mismatch due to the recipe just being "3.14".
(BOOTIMG_NAME_4k, BOOTIMGNAME_2k): Do not interpolate the date and
time into the variable contents, but rather a piece of syntax
which some shell code expand to a date time. This is the same
trick that was applied in the linux-quic recipe under mdm9x28 world,
needed in order to stabilize the hash.
(do_bootimg): Task made to depend on mdm9x15-image-initramfs's
do_image_complete task, rather than do_rootfs. Logic related to
the BOOTIMG_NAME_4k and BOOTIMGNAME_2k DATETIME trick added
here.
(do_tag_config): We have to cd to the right directory for the
filtering step here to work.
(do_oldconfig): Newly invented task: this is how we pull our ready-made
kernel config file into the build. It is staged after do_configure
and before do_tag_config.
(do_kernel_configme): Assert empty task to suppress this piece
of logic that comes from yocto-kernel.inc; that configuration
mechanism isn't working for us here.
(do_install_append): Additional install steps to prepare kernel
headers for the sysroot. This is similar to what is done in
linux-quic, but there are some additional twists because the linux-yocto
kernel build system uses build directory that doesn't contain
the source files; we have to copy some files from ${S}.

* meta-swi-mdm9x15/recipes-kernel/linux/linux-yocto-rt_3.14.bbappend: Renamed
to .bb; added require recipes-kernel/linux/linux-yocto.inc.

Resolves: LXSWIREF-125
Change-Id: I3db86b33d88eabdf27090218093ad315cd2a7237