[sftp-server] Create symbolic link in /usr/libexec
Due to an error in filename, Yocto missed the openssh build step
to create /usr/libexec and add a symlink to sftp-server. Rename
the file so that it could be properly detected during build.

As it turns out, Legato has a dependency on openssl that was
not explicit until now, resulting in a race condition.

Resolves: LXSWIREF-141
Change-Id: I670da07c073631fbc70994ddf154bb26533bd5ca
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>