Scale back our "mdev -s" changes.
The changes done for LXSWIREF-60 cause the breakage which is responsible
for LXSWIREF-142.

The manifestation is that the MTD block devices such as /dev/mtdblock4
were being wrongly created as character devices.

The main reason for this is that inclusion of the /sys/devices tree in
the "mdev -s" scan causes mdevl to find these block devices, but without
the clues which it uses to distinguish them as block devices; it
consequently wrongly installs them as character devices.

A fix for that, without regressing LXSWIREF-60, is to only scan
/sys/devices/platform, since no block devices are found there.

However, when this is done, another flaw is exposed; there is some
problem with the "recursive-action-optimize.patch" which renders the
system unusable: mdev is not properly scanning /sys/class. Somehow,
scanning /sys/devices makes up for it, but scanning only
/sys/devices/platform doesn't.

The recursive optimization patch isn't critically needed because due to
the reduced scope of the /sys/devices/platform scan, "mdev -s" takes
only about 1.3 seconds. There are a 4X fewer file system nodes to visit:

  root@swi-mdm9x15:~# find /sys/devices | wc -l
  17940
  root@swi-mdm9x15:~# find /sys/devices/platform | wc -l
  4795

* common/recipes-core/busybox/busybox_%.bbappend (SRC_URI): Take out the
recursive-action-optimize.patch.

* common/recipes-core/busybox/files/mdev-dev-bus-usb.patch: Alter the
patch so that /sys/devices/platform is scanned, not all of
/sys/devices. This will still find the USB bus nodes, preserving
the fix for LXSWIREF-60.

Resolves: LXSWIREF-142
Change-Id: I8c655a1746a803f3246a599bee727fd3d22519e7