mdm9x28: Improve the robustness of aplay and arec
Issue Description:
After review our previous chagne of aplay and arec file, find some
defect then submit this case to improve it

Investigation:
The access of cleanuping sub-modules is not correct, and it is
complicated to add all the corresponding settings for the code.So, move
all the cleanup features to the pthread_push_cleanup call.

Root Cause:
n/a

Affected platforms:
AR758x

Solution:
force to call cleanup to release all the resources.

Resolves: LXSWIREF-213
Related: QTI9X28-2076
Change-Id: Ia2f9216f87f198c0f4f5e0b3ba680263feba2324
Signed-off-by: Traving Chen <trchen@sierrawireless.com>