[openssh] Remove invalid .bbappend
There is an invalid .bbappend file that has a typo.
The purpose of this file was to provide sftp-server, but it
is already being provided on the latest version of openssh.

As it turns out, Legato has a dependency on openssl that was
not explicit until now, resulting in a race condition.

Resolves: LXSWIREF-141
Change-Id: I670da07c073631fbc70994ddf154bb26533bd5ca
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>